// src/common/types.ts
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  Upload: any
}

export type Address = {
  __typename?: "Address"
  street?: Maybe<Scalars["String"]>
  suite?: Maybe<Scalars["String"]>
  city?: Maybe<Scalars["String"]>
  zipcode?: Maybe<Scalars["String"]>
  geo?: Maybe<Geo>
}
// ...
export type AlbumsPage = {
  __typename?: "AlbumsPage"
  data?: Maybe<Array<Maybe<Album>>>
  links?: Maybe<PaginationLinks>
  meta?: Maybe<PageMetadata>
}
// ...
export type Company = {
  __typename?: "Company"
  name?: Maybe<Scalars["String"]>
  catchPhrase?: Maybe<Scalars["String"]>
  bs?: Maybe<Scalars["String"]>
}
// ...
export type PostsPage = {
  __typename?: "PostsPage"
  data?: Maybe<Array<Maybe<Post>>>
  links?: Maybe<PaginationLinks>
  meta?: Maybe<PageMetadata>
}
// ...
export type TodoQuery = {
  todo: Todo;
}
// ...
export type Todo = {
  __typename?: "Todo"
  id: Scalars["ID"]
  title: Scalars["String"]
  completed: Scalars["Boolean"]
  user: User
}

export type TodosPage = {
  __typename?: "TodosPage"
  data?: Maybe<Array<Maybe<Todo>>>
  links?: Maybe<PaginationLinks>
  meta?: Maybe<PageMetadata>
}
// ...
export type UpdateTodoInput = {
  title?: Maybe<Scalars["String"]>
  completed?: Maybe<Scalars["Boolean"]>
}

export type UpdateUserInput = {
  name?: Maybe<Scalars["String"]>
  username?: Maybe<Scalars["String"]>
  email?: Maybe<Scalars["String"]>
  address?: Maybe<AddressInput>
  phone?: Maybe<Scalars["String"]>
  website?: Maybe<Scalars["String"]>
  company?: Maybe<CompanyInput>
}

export type User = {
  __typename?: "User"
  id?: Maybe<Scalars["ID"]>
  name?: Maybe<Scalars["String"]>
  username?: Maybe<Scalars["String"]>
  email?: Maybe<Scalars["String"]>
  address?: Maybe<Address>
  phone?: Maybe<Scalars["String"]>
  website?: Maybe<Scalars["String"]>
  company?: Maybe<Company>
  posts?: Maybe<PostsPage>
  albums?: Maybe<AlbumsPage>
  todos?: Maybe<TodosPage>
}